RHMAP-11059 Add default location for Geolocation if query fails
Currently when running the weather application in an emulator the
obtaining of a location will fail due to there being no location
provided by the emulated device unless it is set manually.

This change makes the user aware of this information as well as adding
a default value for the location inside the app logic so that even
if the obtaining of a real location fails the weather section of the
app can still be used.

This also includes a small restructuring of the functions to work
better with this updated flow.