RHMAP-11059 Update error message and update `main.js`
This updates the logic for the error messages that allows a custom
instruction to the end of the message instead of just asking the user
to try again later. If one is not provided then it will default to "try
again later".

Also the resulting `main.js` from `grunt browserify` has been included.